Refine rsync target to rsync-eoas and rsync-shelob.
rsync-eoas syncs files to the user's EOAS public_html/ directory.
It can be used from inside or outside the EOAS network.

rsync-shelob syncs files to the salishse sub-domain space on the shelob
web server.
Files associated with the SoG-bloomcast spring diatom bloom prediction
are excluded from the sync to avoid over-writing the files that are
managed by that project.
rsync-shelob can only be used from a machine on the EOAS network.