Implement handling of unavailable biology image loops
Modified the biology image loop handling process to deal with cases where image
loops are not available. This is necessary due to the removal of the ciliates
variable, and the name change of the turbidity variable in V2021-11.

A new list, 'unavailable_loops' has been created to store image loops
that are not available. Those loops are then removed from a copy of the biology
loops data structure to produce the image loops data structure that is passed to
the page renderer.