Update metadata for time series plots
The titles, link texts, and svg names for 2 figures on the SalishSeaCast page
have been updated. The entities represented in the figures have been changed -
from Mesodinium rubrum to Diatoms, and from Mesozooplankton and Microzooplankton
to Z1 and Z2 - to reflect changes in the output fields from the V21-11 model.