Docs maintenance (#96)
* Update ReadTheDocs build config

Upgraded the OS to Ubuntu 24.04 and updated the Python toolchain to
mambaforge-23.11 in the .readthedocs.yaml file. This ensures compatibility with
the latest environment settings available on readthedocs.

* Update Sphinx & optional dependencies version pins

Updated Sphinx and optional dependencies packages pinned versions in
`environment-dev.yaml`, `environment-test.yaml`, and `environment-rtd.yaml`
in preparation for Python 3.13 environment.

* Update Sphinx source_suffix config to mapping

The deprecated list of strings appears to have been removed in Sphinx=8.

* Update Conda documentation URLs

The Conda URLs in the development and installation guides were outdated. This
change updates the URLs to point to the correct and current links under the
official Conda documentation.

* Update docs example output from Sphinx and pytest

Updated Sphinx build and linkcheck commands example outputs. Also revised the
pytest output to reflect the newer plugin versions and test suite logging
changes.