Drop VHFR FVCOM model results image pages (#121)
* Remove VHFR FVCOM results link from navigation menu

The VHFR FVCOM results page link was removed from the navigation menu.
This reflects the fact that we have not been running the VHFR FVCOM model
configurations since 22-Feb-2023. So, there are no results figures to view.

* Remove VHFR FVCOM results routes, views & templates

Deleted all views, templates, and Pyramid routes related to VHFR FVCOM model
results. This reflects the discontinuation of runs of the VHFR FVCOM model
configurations as of 22-Feb-2023, and the absence of relevant results.

* Fix param name in `ImageLoop.filename()` docstring

Corrected the parameter name from `run_date` to `file_date` in the
`ImageLoop.filename()` method's docstring. This ensures the docstring accurately
reflects the method's signature, improving developer clarity.