Fix bugs in GHA deployment workflow (#136)
* Use matrix.python-version in GHA deployment workflow

Replaced `inputs.python-version` with `matrix.python-version` in the
`deployment` workflow. This fixes a copy-paste error from the reusable
`pytest-with-coverage` workflow where the Python version is an input.

* Add date-based cache key to GHA deployment workflow

Introduced a missing step to generate a date-based cache key for downloads in
the `deployment` workflow. This ensures that downloads are cached for 1
day. It is unclear how long the downloads were cached previously when the date
part of the cache key was empty.